analog: noise_source, fastnoise_source: python bindings for [u]int64_t
While we do want to have an uint64_t seed API in C++, Pybind will not
allow us to pass in negative ints that way. If we instead settle on a
wrapper that's int64_t, we only get positive ints up to 2**63-1. To get
up to 2**64-1, we need to also explicitly declare an uint64_t wrapper,
and forbid automatic type conversion.

While figuring this out, I needed to add a bit of debug logging to
fastnoise_source; that does feel like a slight enhancement, so I'm
leaving that in.

Signed-off-by: Marcus Müller <mmueller@gnuradio.org>